// src/typescript/simplestamp/models/simplestamp/v1/operation.ts
export enum OperationType {
  OPERATION_TYPE_UNKNOWN = 0,
  OPERATION_TYPE_ATTESTATION = 1,
  OPERATION_TYPE_SHA1 = 2,
  OPERATION_TYPE_RIPEMD160 = 3,
  OPERATION_TYPE_SHA256 = 8,
  OPERATION_TYPE_APPEND = 240,
  OPERATION_TYPE_PREPEND = 241,
  OPERATION_TYPE_REVERSE = 242,
  OPERATION_TYPE_HEXLIFY = 243,
  UNRECOGNIZED = -1,
}

export interface Operation {
  type: OperationType;
  /**
   * (optional) The value of the argument passed into the operation routine in the
   * case of append and prepend.
   */
  value: Uint8Array;
  /** (optional) An update to the Attestation's status */
  status: AttestationStatus;
  /** The fully qualified URL of the calendar server the stamping was made on. */
  calendarUrl: string;
  /** If the attestation indicates it is on chain */
  blockHeight: number;
}

function createBaseOperation(): Operation {
  return { type: 0, value: new Uint8Array(0), status: 0, calendarUrl: "", blockHeight: 0 };
}

export const Operation: MessageFns<Operation> = {
  encode(message: Operation, writer: BinaryWriter = new BinaryWriter()): BinaryWriter {
    if (message.type !== 0) {
      writer.uint32(8).int32(message.type);
    }
    if (message.value.length !== 0) {
      writer.uint32(18).bytes(message.value);
    }
    if (message.status !== 0) {
      writer.uint32(24).int32(message.status);
    }
    if (message.calendarUrl !== "") {
      writer.uint32(34).string(message.calendarUrl);
    }
    if (message.blockHeight !== 0) {
      writer.uint32(40).uint32(message.blockHeight);
    }
    return writer;
  },

  decode(input: BinaryReader | Uint8Array, length?: number): Operation {
    const reader = input instanceof BinaryReader ? input : new BinaryReader(input);
    const end = length === undefined ? reader.len : reader.pos + length;
    const message = createBaseOperation();
    while (reader.pos < end) {
      const tag = reader.uint32();
      switch (tag >>> 3) {
        case 1: {
          if (tag !== 8) {
            break;
          }

          message.type = reader.int32() as any;
          continue;
        }
        case 2: {
          if (tag !== 18) {
            break;
          }

          message.value = reader.bytes();
          continue;
        }
        case 3: {
          if (tag !== 24) {
            break;
          }

          message.status = reader.int32() as any;
          continue;
        }
        case 4: {
          if (tag !== 34) {
            break;
          }

          message.calendarUrl = reader.string();
          continue;
        }
        case 5: {
          if (tag !== 40) {
            break;
          }

          message.blockHeight = reader.uint32();
          continue;
        }
      }
      if ((tag & 7) === 4 || tag === 0) {
        break;
      }
      reader.skip(tag & 7);
    }
    return message;
  },

  create<I extends Exact<DeepPartial<Operation>, I>>(base?: I): Operation {
    return Operation.fromPartial(base ?? ({} as any));
  },
  fromPartial<I extends Exact<DeepPartial<Operation>, I>>(object: I): Operation {
    const message = createBaseOperation();
    message.type = object.type ?? 0;
    message.value = object.value ?? new Uint8Array(0);
    message.status = object.status ?? 0;
    message.calendarUrl = object.calendarUrl ?? "";
    message.blockHeight = object.blockHeight ?? 0;
    return message;
  },
};

type Builtin = Date | Function | Uint8Array | string | number | boolean | undefined;

export type DeepPartial<T> = T extends Builtin ? T
  : T extends globalThis.Array<infer U> ? globalThis.Array<DeepPartial<U>>
  : T extends ReadonlyArray<infer U> ? ReadonlyArray<DeepPartial<U>>
  : T extends {} ? { [K in keyof T]?: DeepPartial<T[K]> }
  : Partial<T>;

type KeysOfUnion<T> = T extends T ? keyof T : never;
export type Exact<P, I extends P> = P extends Builtin ? P
  : P & { [K in keyof P]: Exact<P[K], I[K]> } & { [K in Exclude<keyof I, KeysOfUnion<P>>]: never };

export interface MessageFns<T> {
  encode(message: T, writer?: BinaryWriter): BinaryWriter;
  decode(input: BinaryReader | Uint8Array, length?: number): T;
  create<I extends Exact<DeepPartial<T>, I>>(base?: I): T;
  fromPartial<I extends Exact<DeepPartial<T>, I>>(object: I): T;
}